Update Different_Pixels_SV.py
This program manipulates the pixels of an image. The user can choose to create a rectangle, a right triangle, or a circle in the program, as well as creating all three shapes on the image. For each of these shapes, the user has to option to change the color or leave it at the default, which is black. In the program, the defaults for the command line arguments are set so that none of the programs will run unless the user asks it to run. For example, the user must specify the center and radius of the circle, or the circle will not be created. For the circle, the user has the option to manipulate the image so that the circle moves. The program is set so that when this option is chosen, the circle will bounce around at random at the edges of the image provided.