Frame_Detection_SV- Description
This program takes in two frames provided by the user. It will find the difference between the images, making all pixels that are the same between the images white and finding the pixel value difference between the frames. It will also eliminate small sections of different pixels to emphasize the main differences between the frames.