Video_Differences_SV- Description
This program inputs a video and pulls out four frames, based on an input of the starting frame. The program first derives A-level images, which are the differences between specific frames. The program goes a step further, creating B-level images, which are the differences between A-level images. Lastly, the program derives C-level images, which are the differences between B-level images. This program compiles a set of created frames that can be used to analyze the differences between frames in a video.