Rotating_Images_SV- Description
This program rotates inputted images. It takes in two different images and rotates them. For one of the images, the user gets to input the coordinates that the image rotates around and chooses the angle that the image will rotate. The user also gets to resize the image as long as the aspect ratio is constant. The second image rotates as an animation, switching frames after a set amount of time provided by the user.