Resizing_Images_SV- Description
This program takes in a number of inputs, including an image, and resizes it. The default in the program is to run the animation section, as denoted by the variable "move". Alternatively, the program can create a single image with set pixel dimensions by choosing the "change" section. A section is chosen by inputting the number 1 and not chosen by the number 0. The program has a number of defaults that will run the program correctly, but the user can change these inputs to cause different features of the program to activate and create different results.