Cropped_Images_SV- Description
This program takes in inputs for an image as well as the height and width of the cropped section. The program will then run an infinite loop where it continually moves, causing the cropped section to constantly change. The program sends a message in the console every time it starts on a new column or a new iteration.