Masking_Images_SV- Description
This program takes in an image as input, as well as the radius of a circle and the speed of the moving circle. With these inputs, the program will create a circular mask and apply it to the image, causing a circular section of the image to be highlighted while the rest of the image is black. The mask will continually move in both directions, allowing the user to view the entire image.