Manipulating_Channels_SV- Description
This program takes in an image as input and then splits the image into different Red-Green-Blue channels. The program then displays seven different images, six of which are filters of the image. The included filters are Red, Blue, Green, Magenta, Yellow, and Cyan. The original image is also displayed for comparison.